do not overwrite ppid if it was successfully parsed from received data (#42)
Co-authored-by: Dorian Scholz <scholz@energy-robotics.com>

## What is the current behavior you want to change?
Kurento overwrites the PPID of data channel packets marking every packet to be of type "String" even when they were sent marked as "Binary".

## What is the new behavior provided by this change?
This PR fixes the code that already tried to preserve the PPID of the SCTP source packet, but eventually overwrote it anyways.

## How has this been tested?
A simple data channel connection from Chrome browser to Chrome browser via Kurento previously failed to receive an ArrayBuffer object, if the ArrayBuffer contains any bytes with the most significant bit set (e.g. `new Uint8Array([255, 128]).buffer`). This is due the PPID set by Kurento to String type, so Chrome tries to parse the data as UTF8 string, which fails.
The change of the PPID can be seen in the Chrome debug logs when run with `--enable-logging --v=2`: packets sent with PPID 0x35 (binary) are turned into PPID 0x33 (string).
After this fix the PPID is preserved by Kurento when sending binary or string data.

Fixes https://github.com/Kurento/bugtracker/issues/631