set-versions: Drop usage of Maven commands, edit pom.xml directly instead
This script does not use the Maven `versions` plugin (with goals such as
`versions:update-parent`, `versions:update-child-modules`, or
`versions:set`) because running Maven requires that the *current*
versions are all correct and existing (available for download or
installed locally).

We have frequently found that this is a limitation, because some times
it is needed to update from an unexisting version (like if some
component is skipping a patch number, during separate development of
different modules), or when doing a Release (when the release version is
not yet available).

It ends up being less troublesome to just edit the pom.xml directly.