cmake: Use generic_find instead of pkg_check_modules to fix includes
Prefer using generic_find() to load our own dependencies (kms-core),
because if those are loaded via pkg_check_modules() then the include
path will be "/usr/include/kurento", which is a bad thing if we want to
make a self-contained build.

For example, if order of virtual methods have changed in local code,
then compiling against headers of an older version will break the
linking of library files.