js: Move JavaScript templates from client to module-creator
Move JavaScript client templates to module-creator/src/main/templates/

This fixes a circular dependency issue between components:

- Kurento Module Creator needs the client templates in order to build
  any Kurento module, including core, elements, filters.
  Given this, it would follow that the JS client is a build dependency
  of the core/elements/filters. However...

- Kurento core/elements/filters are a dependency of the JS client.

- Thus it was impossible to make a from-scratch, clean slate build.

The dirty solution was that Module Creator did a 'git clone' of the
JS client repo, in order to access its templates! This of course means
that it was impossible to develop the JS client, and test the changes
locally, while at the same time trying out the module generation with
your local changes.

Not that it matters a lot now that development of the client is very
minimal, but still, it hindered maintenance of this module.

The implemented solution seems better: Module Creator, as the only user
of these templates, could just contain them. It's not ideal because
the templates technically belong to the JS client, but it's immensely
better than doing a 'git clone' in the spot...