server: Use random ports for SCTP (WebRTC DataChannels)
Use a random port for SCTP-over-DTLS (as used in WebRTC).

Before, the port 5000 was always used, but this causes a lot of
`SCTP_CANT_STR_ASSOC` errors. With random ports, the error still happens
but seems to be much less frequent.

See bug report: https://github.com/sctplab/usrsctp/issues/314