server: Rename all filters to "kms*" to avoid conflicts with GStreamer
GStreamer docs recommend to use "a unique name of the plugin (ideally
prefixed with an application- or library-specific namespace prefix in
order to avoid name conflicts in case a similar plugin with the same
name ever gets added to GStreamer)":

https://gstreamer.freedesktop.org/documentation/gstreamer/gstplugin.html#gst_plugin_register_static

This actually happened, and the Kurento plugin "faceoverlay" happened to
conflict with the GStreamer plugin of the same name. So let's rename all
built-in plugin names to start with "kms".

Also, lots of vars have been renamed to "self", to follow with the
coding conventions of the rest of the media server.

Lastly, the helper class "Classifier" has been removed from the
KmsFaceDetector filter; instead, plain OpenCV classifier instance is
used directly, just like done in the DataChannelExample module.