Avoid race conditions on UUID generation (#4)
Avoid race conditions on UUID generation that in turn may have a big impact on generating duplicated object Id for different Kurento objects.

## What is the current behavior you want to change?
This change is basically a bugfix to avoid race condition on UUID generation that in turn is used on generating unique object references.

We have observed in production a race condition when creating simultaneously two MediaPipelines. The outcome of those race conditions is that both pipelines get associated to the same Id, thus making it impossible to the client application to later on create MediaElements associated to the expected MediaPipeline, as both pipelines are using the same Id.

We have been able to trace this issue to the UUID generation that is part of the object Id. It seems that in certain concurrency conditions  boost UUID can generate duplicated ids. We have even checked comments on boost UUID on several forums (e.g. https://github.com/boostorg/uuid/issues/121), and verified that boost UUID is not mt safe. And in the UUIDGenerator class, it was used without any multithreading protection.



## What is the new behavior provided by this change?
We have just implemented a Mutex that prevents boost UUID generator to be used concurrently by more than one thread. This change shoudl not hit performance as UUID generation is only used on Kurento object creations and the mutex is held just for the UUID generation, so not much contention should be caused.

## How has this been tested?
IT has been tested in production and on test suite that reproduced the issue and after applying this PR issue was fixed. We haven't included test suite because it includes certain elements of client application where issue was first observed.