server: GenericFind.cmake: Remove redundant call to find_package()
At this point in the build script, find_package(${GF_LIBNAME}) had been
already called in the previous code block, so it doesn't seem that it
was necessary to call it again here.

The log output of calling this function was seen twice in the console,
so it appears as if it was actually being called unnecessarily.