server: Do not update dependency on Kurento Media Server
Don't be too happy to update the version requirement of Kurento Module
Creator. Older versions should be able to generate code for new ones.
This line is commented out for reference, but should be run manually
only on those cases where a breaking change has been introduced in the
module creator and the upgraded requirement is actually a necessity.