server: Fix incorrect out caps from audiomixer element
The audiomixer was _sometimes_ outputting
`audio/x-raw, channels=(int)1`
after the downstream WEBM muxer (in RecorderEndpoint) had been
configured for 2 channels.

This caused an error in the webmmuxer, because changing caps is not
supported.

The exact reason for this sudden change wasn't found, but it seems that
forcing the output caps of audiomixer is enough to make it behave
correctly and not change its output caps in the middle of processing.