clients/java: Disable building jsonrpc-demo-server by default
Use the "demo" profile instead.

The jsonrpc-demo-server package takes 50MB because it is self-contained
and includes all of its dependencies. That's fine for running a quick
demo but not good for CI builds that copy the artifacts around and this
package size was dominating the overall workspace usage of the builds.