doc: Fix anchor links to latest bug-fix release in Release Notes
Fix anchor link to the latest bug-fix release, because dots in anchors
get converted into bars, so the title "v7.0.1" becomes "#v7_0_1". Also,
for some reason, anchors cannot start with a number but only letters, so
"7.0.1" becomes "#id1" instead of "#7_0_1".

For all these reasons, we give up on trying to link to the exact section
of the latest bug-fix release, and instead just link to the global
bug-fix releases section.