server: RecorderEndpoint/UriEndpoint parse username/password from URLs
Use GstUri to parse the destination URL and extract the
username:password field (called "userinfo").

Then, split the userinfo from the colon (:).

Lastly, URL-decode each of the resulting parts.

NOTE: GStreamer 1.16 (Ubuntu 20.04) does NOT support extracting a
URL-encoded userinfo field. This means that the username and password
will be already decoded, and the splitting on ":" will happen at the
first colon found, which won't work if the username itself had a colon.

So, until we update to a newer Ubuntu/GStreamer version,
THE USERNAME CANNOT CONTAIN A COLON.

Also delete the unneeded function kms_is_valid_uri()`. Better to
delegate that check to the GstUri class constructor, which validates the
URI and ecides if it is valid or not.