client/java: Don't forget sessionId if server response's is null
This happened in case of server error response: the sessionId is not set
by the server in the error messages, so sessionId field would be null.
Then, a null value would be set here, which meant that the next request
from this client would be sent without a sessionId.

The JavaScript client doesn't behave like that, and instead it still
uses the same sessionId in the next request after a server error.