.github: Workflow file renames that were pending commit
Some normalization of the files, to follow docs on Reusable Commits,
which call them "caller" and "called" workflows.

Also, closer follow of the same directory path in the workflow names.
So, the workflow for "clients/java" is called exactly like that, and
this applies to all workflows. Exception for javascript: "js" is used as
a shorthand.

Also no need to use a generic word like "build" in names. It doesn't say
anything we didn't know, of course most CI jobs are for "building"
things.