ci: Revert "maven_deploy: use the ci-build local repository for deployments too"
Partial revert of e3d230059ef7ae4ce914652c0984242083232126.

Reason: The "ci-build" Maven profile should be used only to deploy into
a local dir, which then later is archived by the CI job. It should not
be used as destination local repository for other tasks, such as the
"mvn install" that is done in maven_deploy_github.