ci: Revert "Rename profile 'deploy-local' to 'ci-build'"
Partial revert of 51725c9593fc1eb80b5654b531047b47d81fb635.

Reason: The name "deploy-local" is a natural extension of the other
deployment profile, "deploy". It also is clearer to signify that this
profile should be used just for deployments to a local repository.

The previous name, "ci-build", could make you think that it should be
enabled for any kind of builds in CI, which is not correct. This profile
is to be used only for running "mvn deploy" in order to deploy into a
local directory, which later gets archived by the CI job.