clients/java: Flexible dependency range from maven-plugin to module-creator
Take whichever version of module-creator is available.

module-creator can evolve faster than maven-plugin, so we don't have
an exact version that can be put in this dependency.

For example: if this was maven-plugin 7.0.1-SNAPSHOT, we cannot assume
that module-creator will also be 7.0.1-SNAPSHOT, because it might
already be on 7.0.2-SNAPSHOT or higher.