clients/java: Flexible dependency range from client to server modules
Take whichever version of server modules are available.

The server modules can evolve faster than the client, so we don't have
an exact version that can be put in these dependencies.

For example: if this was client 7.0.1-SNAPSHOT, we cannot assume
that kms-api-core will also be 7.0.1-SNAPSHOT, because it might
already be on 7.0.2-SNAPSHOT or higher.